requirements.txt
certifi==2018.1.18
chardet==3.0.4
humanize==0.5.1
idna==2.6
python-dateutil==2.6.1
requests==2.18.4
six==1.11.0
urllib3==1.22
